Remove $LastChangedDate$ from usage
Usage printout contained a row with literal content
'$LastChangedDate$.'. It looks like a tag that should be replaced with
a timestamp, but such thing does not happen and the tag is displayed
as-is. Remove it.

Another option would be to actually insert the last changed date into
that position. It would be easy to do, as there are already other items
in usage, such as program name and version number, that are inserted.
But that would require maintaining the correct value, so it is better
to just remove the tag.